requirements.txt
Django==2.2.1
djangorestframework==3.9.3
sqlparse==0.3.0
django-bootstrap==0.2.4
django-bootstrap-navbar==0.0.13
django-bootstrap3==11.0.0
django-crispy-forms==1.7.2
django-tables2==2.0.6
pkg-resources==0.0.0
pytz==2019.1
sqlparse==0.3.0
